user-actor-my-pets - move list-my-pets-handler into lib/handlers
